Many fixes
Fixed the A logo to have some margin to the top of the page
Fixed the left side of the navbar as it caused some separation for the borders
I kept my text in the navbar left aligned as it looks better to me
Projects cards and the contact information wasn’t vertically aligned for big screens. I fixed it
Fixed the image issue in SCS page
I added the right-border of the navbar as I have missed it
I changed the back to top button color and made the arrow a link too
I decided to have the hamburger menu on top for small screens and move it to the middle of the screen for medium ones before it disappears but does not seem to make sense to be that way. I changed it back to be on top left
I added some info and links to the about-me part
I fixed the problem with the menu button on small screens. I have made a mistake in the styling for it in the SCS page only. Works fine on all other pages and on the SCS page too now.
My hero image was not centred correctly. I have tried with different attributes for the different screen sizes but it does not seem good enough. I decided to look and see why it does follow the screen, not the hero area/container. It turns out that no ancestor element got any positioning. So I did add a relative position to its parent element in order to make it aligned with it. It works fine now and seems in the centre of all screens regardless of the navbar.
I had some issues with the position of the scroll down button as it did not seem to follow the correct element. I did some changes and made it follow the hero image as its parent element together with the hero text and now they flow together on any screen size
Fixed the scroll button in the navbar on small height screens as it was below the hero image
Fixed the submit button to be centred