Fix hamburger click issue
There was an issue that the hamburger button on small screens click led to the page to go to the top. It all worked fine but not great user experience if they have to scroll back down every time they click on the menu button and close it after that.