Fixed issues
I managed to fix the big issue I was having with the successfully submitted form, where users could refresh the page and the form gets submitted again and again.
I fixed the issue that I was having on small screens. The error message was out of sight. I added some JS to make the viewport scroll back up to the message on submit if error occurs
Fixed some small styling issues with the form on some screen sizes
Removed the success message on form submission after 8s so that the user is not confused if they want to send another message