Feedback received
I received feedback about my page and implemented some changes:
Made the hamburger menu icon bigger
Fixed the JS error message to appear correctly on small screens
Changed some code to make the success message disappear if the user decided to submit the form again with new data
Added green border if the input is correct and checked with submit button, they used to be red - misleading
Changed the submit button to reflect the page colours and made some other changes in the overall styling
Had an issue with the textarea. I used input previously and it worked fine but had to change it as it does not allow more than 1 line and it is not the correct way to use it for long text. Once I did that the user input got deleted if errors occurs and placeholder was not visible. I did change some stuff after reading on stackoverflow
I added alphanumeric Regex to my JS validation function to make sure white spaces are not allowed