More bugs fixed
I made the hamburger small again as it did not look nice when bigger. I checked on my phone and a few other phones and it looks better when smaller
I had an issue with the success message staying all the time after the changes I did last time. I managed to get that fixed and now it disappears if the user starts the form again and error is thrown
I removed the hamburger on small screens when the navbar is open and get it back when closed. Also made sure that changing the screen width would not affect the behaviour in any way