Feedback received
I worked on the feedback I received about my project:
I added meta description to the page.
I made the hamburger menu a bit bigger for small screens as it was pointed out. I did have this feedback before and decided that it is good and not too small as I tried it on different actual phones. Now I increased the size a little bit only as otherwise it just becomes too ugly.
I added border to the input elements to be same colour as the background when not on focus. That way it will have a border and it will just change colour to red when not valid info is typed. That way I avoid the change in the size of the input elements I was struggling with. Thanks a lot for the tip to Kayleigh.
I did have some kind of margin between the logo and the navbar. I removed that and made sure that the text is aligned with the logo and no big margin between them.
I made the logo centred for small screens when the sidebar is visible. It was a bit to the right
Fixed a title padding issue for some screen sizes. The title was not aligned with the rest of the text in the project cards
My project cards were not centred for the biggest screen. I think I might have missed that because my screen is exactly the size where it looks centred. I added margin-left and right auto to make the container centred for any screen bigger than XXL.
I styled the scrollbar in the tooltip as it was the default one. I made it round and changed the colour.